Adjust tests calculating subtotal
